+++ app/mailers/notification_mailer.rb
@@ -101,7 +101,10 @@ class NotificationMailer < ActionMailer::Base
     else
       @locale = Rails.configuration.i18n.default_locale
     end
-    title = I18n::translate(:new_reply, locale: @locale) + ': ' + reply.ticket.subject
+    title =  reply.ticket.subject
+    if user.agent?
+      title = I18n::translate(:new_reply, locale: @locale) + ': ' + title
+    end
 
     add_attachments(reply)
     add_reference_message_ids(reply)